switch datepicker from react-date-picker to react-datetime.
