--- app/javascript/components/needsPicker/Picker.scss
@@ -1,120 +1,0 @@
-@import '~foundation-sites/scss/util/util';
-.DayPickerInput, .conditional-time-select {
-  width: 100%;
-}
-.DayPicker-Month {
-  margin-left: 0;
-  margin-right: 0;
-}
-
-.DayPicker-Day--today {
-  color: #2FCFD9;
-}
-
-.DayPicker-Day--highlighted {
-  position: relative;
-  &::after{
-    content: '';
-    display: inline-block;
-    width: 4px;
-    height: 4px;
-    -moz-border-radius: 7.5px;
-    -webkit-border-radius: 7.5px;
-    border-radius: 7.5px;
-    background-color: #e86071;
-    display: block;
-    position: absolute;
-    left: calc(50% - 2px);
-  }
-}
-.DayPicker--interactionDisabled {
-  display: flex;
-  justify-content: center;
-}
-
-.blockout-modal-inner-content {
-  .DayPickerInput-Overlay {
-    left: unset;
-    right: 0;
-  }
-}
-
-.blockout-day-heading {
-  font-weight: bold;
-  margin-top: 1rem;
-  margin-bottom: .5rem;
-}
-
-.blockout-container {
-  margin-bottom: 1rem;
-  cursor: pointer;
-  .card-divider, .card-section {
-    padding: .5rem;
-  }
-}
-
-.blockout-time {
-  font-size: .7rem;
-}
-
-.blockout-modal-overlay {
-  display: block !important;
-}
-
-.blockout-modal-content {
-  display: block !important;
-  max-width: 100vw;
-  outline: none;
-  padding: 0;
-  border: none;
-  @include breakpoint(medium) {
-    overflow: unset;
-  }
-}
-
-.blockout-modal-header {
-  background-color: #2FCFD9;
-  padding: 1rem;
-  color: #fff;
-  font-weight: bold;
-}
-
-.blockout-modal-inner-content {
-  padding: 1rem;
-  hr {
-    width: calc(100% + 2rem);
-    margin: 0 0 1rem -1rem;
-  }
-}
-
-.blockout-modal-footer {
-  border-top: 1px solid #ccc;
-  padding: .8rem;
-  .button {
-    margin-bottom: 0;
-  }
-}
-
-.blockout-repeat-inputs {
-  .cell {
-    margin-right: .5rem;
-  }
-}
-
-.blockout-edit-recurrence-options {
-  button {
-    &.primary {
-      .subtext {
-        color: #d5fcff;
-      }
-    }
-    .subtext {
-      margin-top: .25rem;
-      font-size: .75rem;
-    }
-  }
-}
-
-.add-new-blockout-button {
-  width: 100%;
-}
