--- app/javascript/components/needsPicker/contexts/picker.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-export default React.createContext({
-  calendarMonth: null,
-  setCalendarMonth: _ => {},
-  authenticityToken: ""
-});
