+++ app/views/needs/_form_fields.html.haml
@@ -11,7 +11,7 @@
   %legend Start Date/Time
   .grid-x
     .cell
-      = react_component("needsPicker/addNeedsPicker/index", { authenticity_token: form_authenticity_token }, {id: 'scheduler'})
+      = react_component("NewNeedDatePicker/index")
 %fieldset.fieldset
   %legend Expected Duration
   .grid-x.grid-margin-x
